container/app/requirements.txt
psycopg2-binary==2.9.*
boto3>=1.*